+++ src/Plainion.GraphViz.Modules.CodeInspection/Inheritance/Actors/InheritanceActor.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Akka.Actor;
 using Plainion.GraphViz.Modules.CodeInspection.Common.Actors;
+using Plainion.GraphViz.Modules.CodeInspection.Common.Analyzers;
 using Plainion.GraphViz.Modules.CodeInspection.Inheritance.Analyzers;
 
 namespace Plainion.GraphViz.Modules.CodeInspection.Inheritance.Actors
@@ -19,9 +20,12 @@ namespace Plainion.GraphViz.Modules.CodeInspection.Inheritance.Actors
 
                 Task.Run<TypeRelationshipDocument>(() =>
                 {
-                    var analyzer = new InheritanceAnalyzer();
-                    analyzer.IgnoreDotNetTypes = r.IgnoreDotNetTypes;
-                    return analyzer.Execute(r.AssemblyLocation, r.TypeToAnalyze, CancellationToken);
+                    using (var resolver = new AssemblyResolver())
+                    {
+                        var analyzer = new InheritanceAnalyzer();
+                        analyzer.IgnoreDotNetTypes = r.IgnoreDotNetTypes;
+                        return analyzer.Execute(r.AssemblyLocation, r.TypeToAnalyze, CancellationToken);
+                    }
                 }, CancellationToken)
                 .ContinueWith<object>(x =>
                 {